Move getLabel from vulcan:intl to vulcan:lib
This will allow projects to replace vulcan:intl and use the full react-intl
