+++ packages/vulcan-forms/test/components.test.js
@@ -161,7 +161,6 @@ const Addresses = createCollection({
 describe('vulcan-forms/components', function() {
   const context = {
     intl: {
-      formatLabel: () => '',
       formatMessage: () => '',
       formatDate: () => '',
       formatTime: () => '',
