+++ packages/vulcan-i18n/lib/modules/shape.js
@@ -29,7 +29,6 @@ export const intlFormatPropTypes = {
   formatNumber: funcReq,
   formatPlural: funcReq,
   formatMessage: funcReq,
-  formatLabel: funcReq,
   formatHTMLMessage: funcReq,
 };
